// web/src/shared/api/generated/models/searchResult.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * NOVA API
 * NOVA provides Network Oriented Visualized Analysis via intuitive network graphs to uncover hidden insights.
 * OpenAPI spec version: 0.1.0
 */

export interface SearchResult {
  /** ページタイトル */
  title: string;
  /** ページURL */
  url: string;
  /** ページ概要 */
  snippet: string;
  /**
   * 記事作成日（オプショナル）
   * @nullable
   */
  date?: string | null;
  /**
   * 最終更新日（オプショナル）
   * @nullable
   */
  last_updated?: string | null;
}